--- rustfmt.toml
@@ -1,1 +1,0 @@
-edition = "2018"
