+++ src/commands.rs
@@ -1,10 +1,9 @@
 //! SMTP commands
 
-use crate::smtp::authentication::{Credentials, Mechanism};
-use crate::smtp::error::Error;
-use crate::smtp::extension::ClientId;
-use crate::smtp::extension::{MailParameter, RcptParameter};
-use crate::smtp::response::Response;
+use crate::authentication::{Credentials, Mechanism};
+use crate::error::Error;
+use crate::extension::{ClientId, MailParameter, RcptParameter};
+use crate::response::Response;
 use crate::EmailAddress;
 use log::debug;
 use std::convert::AsRef;
@@ -12,10 +11,6 @@ use std::fmt::{self, Display, Formatter};
 
 /// EHLO command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct EhloCommand {
     client_id: ClientId,
 }
@@ -35,10 +30,6 @@ impl EhloCommand {
 
 /// STARTTLS command
 #[derive(PartialEq, Eq, Clone, Debug, Copy)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct StarttlsCommand;
 
 impl Display for StarttlsCommand {
@@ -49,10 +40,6 @@ impl Display for StarttlsCommand {
 
 /// MAIL command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct MailCommand {
     sender: Option<EmailAddress>,
     parameters: Vec<MailParameter>,
@@ -81,10 +68,6 @@ impl MailCommand {
 
 /// RCPT command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct RcptCommand {
     recipient: EmailAddress,
     parameters: Vec<RcptParameter>,
@@ -112,10 +95,6 @@ impl RcptCommand {
 
 /// DATA command
 #[derive(PartialEq, Eq, Clone, Debug, Copy)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct DataCommand;
 
 impl Display for DataCommand {
@@ -126,10 +105,6 @@ impl Display for DataCommand {
 
 /// QUIT command
 #[derive(PartialEq, Eq, Clone, Debug, Copy)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct QuitCommand;
 
 impl Display for QuitCommand {
@@ -140,10 +115,6 @@ impl Display for QuitCommand {
 
 /// NOOP command
 #[derive(PartialEq, Eq, Clone, Debug, Copy)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct NoopCommand;
 
 impl Display for NoopCommand {
@@ -154,10 +125,6 @@ impl Display for NoopCommand {
 
 /// HELP command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct HelpCommand {
     argument: Option<String>,
 }
@@ -181,10 +148,6 @@ impl HelpCommand {
 
 /// VRFY command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct VrfyCommand {
     argument: String,
 }
@@ -204,10 +167,6 @@ impl VrfyCommand {
 
 /// EXPN command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct ExpnCommand {
     argument: String,
 }
@@ -227,10 +186,6 @@ impl ExpnCommand {
 
 /// RSET command
 #[derive(PartialEq, Eq, Clone, Debug, Copy)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct RsetCommand;
 
 impl Display for RsetCommand {
@@ -241,10 +196,6 @@ impl Display for RsetCommand {
 
 /// AUTH command
 #[derive(PartialEq, Eq, Clone, Debug)]
-#[cfg_attr(
-    feature = "serde-impls",
-    derive(serde_derive::Serialize, serde_derive::Deserialize)
-)]
 pub struct AuthCommand {
     mechanism: Mechanism,
     credentials: Credentials,
@@ -329,7 +280,7 @@ impl AuthCommand {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::smtp::extension::MailBodyParameter;
+    use crate::extension::MailBodyParameter;
 
     #[test]
     fn test_display() {
