--- src/smtp/client/mod.rs
@@ -1,9 +1,0 @@
-//! SMTP client
-
-mod codec;
-mod inner;
-pub mod mock;
-pub mod net;
-
-pub use self::codec::*;
-pub use self::inner::*;
